model/pyproject.toml
[tool.poetry]
name = "caddy-model"
version = "0.1.0"
description = "Caddy Model API."
authors = ["Alex Moore <alex.moore@cabinetoffice.gov.uk>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^0.2.9"
python-dotenv = "^1.0.1"
langchain = "^0.3.7"
fastapi = "^0.115.5"
pydantic = "^2.10.1"
langchain-aws = "^0.2.7"
langchain-community = "^0.3.7"
boto3 = "^1.35.68"
requests-aws4auth = "^1.3.1"
opensearch-py = "^2.7.1"
uvicorn = "^0.32.1"
streamlit = "^1.40.1"
tiktoken = "^0.8.0"
mcp = {extras = ["cli"], version = "==1.8.0"}
python-multipart = "^0.0.20"
markitdown = {extras = ["docx", "outlook", "pdf", "pptx", "xls", "xlsx"], version = "^0.1.1"}
tqdm = "^4.67.1"
click = "^8.1.8"
pytest-env = "^1.1.5"
sentry-sdk = {extras = ["fastapi", "starlette"], version = "^2.29.1"}
multidict = "==6.4.2"
alembic = "^1.15.2"
sqlmodel = "^0.0.24"
psycopg2-binary = "^2.9.5"
pytest-postgresql = "^7.0.2"
email-validator = "^2.2.0"
pyjwt = "^2.9.0"
python-keycloak = "^4.7.3"
cryptography = "^44.0.1"
pgvector = "^0.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
ruff = "^0.8.0"
pytest-cov = "^6.0.0"
pre-commit = "^4.0.1"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
env = [
  "ENVIRONMENT=test"
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
]
ignore = ["E501"]  # line length violations

[tool.poetry.scripts]
upload-files = "model.scripts.upload_files:ingest_files"
